+++ src/luce/birra/PriceActivity.java
@@ -140,14 +140,14 @@ public class PriceActivity extends FragmentActivity implements LoaderCallbacks<C
     					DialogScreen getkol = new DialogScreen(PriceActivity.this,PriceActivity.this,R.id.cb_Kol_Price)
     							 .setDialogScreenListener(new DialogListener() {
     								@Override
-    								public void OnSelectedKol(float k) {
+    								public void OnSelectedKol(double k) {
     									if (k!=0) 
     									{
     										long countT=0;
     				    					Cursor cc = MainActivity.db.getRawData ("select id_tmc, ed, id_post, price from tmc_price where _id="+idd,null);
     				    					   if (cc.moveToFirst()) { 
     				    					        do {countT=1;
-    				    					        	MainActivity.db.updRec("tmc_price", idd, "price", k);
+    				    					        	MainActivity.db.updRec("tmc_price", idd, "price", (float)k);
     				    					        		//MainActivity.db.addRecRASXODcount(cc.getInt(cc.getColumnIndex("id_tmc")), cc.getFloat(cc.getColumnIndex("kol")), (byte)cc.getInt(cc.getColumnIndex("ed")), 0,0, cc.getInt(cc.getColumnIndex("id_post")), 0, "обнуление остатка id="+idd, MainActivity.getIntDataTime(), 1);
     				    					        		//MainActivity.db.addRecPRIXOD(cc.getInt(cc.getColumnIndex("id_tmc")), k, (byte)cc.getInt(cc.getColumnIndex("ed")), 0,0, cc.getInt(cc.getColumnIndex("id_post")), "изменение остатка id="+idd, MainActivity.getIntDataTime(), (byte)1);
     				    					        } while (cc.moveToNext());
